fix: migrate /api/book/event to trpc route viewer.bookings.event
