+++ apps/web/calendso.yaml
@@ -277,74 +277,6 @@ paths:
                   type: string
                 integration:
                   type: string
-  "/api/book/event":
-    post:
-      description: Creates a booking in the user's calendar.
-      summary: Creates a booking for a user
-      tags:
-        - Booking
-      parameters: []
-      requestBody:
-        content:
-          application/json:
-            schema:
-              type: object
-              properties:
-                start:
-                  type: string
-                end:
-                  type: string
-                rescheduleUid:
-                  type: string
-                eventTypeId:
-                  type: string
-                email:
-                  type: string
-                name:
-                  type: string
-                timeZone:
-                  type: string
-                guests:
-                  type: array
-                  items: {}
-                users:
-                  type: array
-                  items: {}
-                user:
-                  type: string
-                notes:
-                  type: string
-                location:
-                  type: string
-                paymentUid:
-                  type: string
-      responses:
-        "204":
-          description: No Content
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  message:
-                    type: string
-        "500":
-          description: Internal Server Error
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  message:
-                    type: string
-              examples: {}
-    parameters:
-      - schema:
-          type: string
-        name: user
-        in: path
-        required: true
-        description: The user whom you wish to book
   /api/book/confirm:
     post:
       description: Accepts an opt-in booking.
